improvement: using a sanitize file
