+++ services/auth/lib/handlers.js
@@ -10,7 +10,6 @@ const { deleteExpiredTokens } = require('./tokens');
 const { getUsersCollectionName } = require('./modules/collectionNames');
 const createObjectId = require('../../../lib/modules/createObjectId');
 const disposableDomains = require('disposable-email-domains');
-const sanitizeHtml = require('sanitize-html');
 
 async function tokenMaintenance(req, res) {
   if (!req?.user?.isAdmin) {
@@ -589,28 +588,6 @@ function isEmailValid(email) {
   );
 }
 
-/**
- * The function `sanitizeHTMLFromXSS` is used to sanitize HTML strings and objects from potential cross-site scripting (XSS) attacks.
- * @param obj - The `obj` parameter is the input object that you want to sanitize from potential XSS (Cross-Site Scripting) attacks.
- * It can be a string or an object containing strings.
- * @returns The function `sanitizeHTMLFromXSS` returns the sanitized version of the input object, with any potentially harmful HTML
- * tags or attributes removed.
- */
-function sanitizeHTMLFromXSS(obj) {
-  if (!obj) return obj;
-  switch (typeof obj) {
-    case 'string':
-      return sanitizeHtml(obj);
-    case 'object':
-      return Object.entries(obj).reduce((acc, [key, value]) => {
-        acc[key] = sanitizeHTMLFromXSS(value);
-        return acc;
-      }, obj);
-    default:
-      return obj;
-  }
-}
-
 async function getUserByInvitationToken(req, res, next) {
   const invitationToken = req.params.invitationToken;
   if (!invitationToken) {
@@ -635,7 +612,6 @@ module.exports = {
   redirectWithError,
   callback,
   getProviders,
-  sanitizeHTMLFromXSS,
   getUsers,
   getAccessTokenForUser,
   me,
